refactor: move test-mapping config schema into homeboy-audit-contract (#8425)
The last audit config-type cascade in extension. TestMappingConfig (and the
BehaviorScenarioNames, IncludeWrapperPolicy, TestVacuityPolicy, PackageNameSource
types it composes, plus the default_test_prefix serde helper) describe an
extension's source->test conventions for audit's structural test-coverage and
test-vacuity detectors. They lived in extension/manifest.rs but are audit's
schema.

Move the whole 5-type closure (185 lines, contiguous) into
homeboy-audit-contract as a `test_mapping` module. BehaviorScenarioNames' impl
uses homeboy_engine_primitives::language::Language::builtin_*, so the contract
gains a homeboy-engine-primitives dependency (no cycle — primitives is lower).
manifest.rs pub-uses them back so extension::TestMappingConfig etc still resolve;
code_audit imports them from homeboy_audit_contract:: directly.

Result: code_audit -> extension real edges drop 12 -> 4, and those 4 are
test-only fixtures (save_manifest, bench, ExtensionManifest) plus one prod type
(ExtensionPhaseTiming). The audit config/type coupling to extension is
essentially gone.

Verified: contract compiles + 16 tests pass; core compiles clean; cargo check
--workspace --tests green.

Refs #8425
