+++ crates/homeboy-audit-contract/Cargo.toml
@@ -13,5 +13,6 @@ name = "homeboy_audit_contract"
 path = "src/lib.rs"
 
 [dependencies]
+homeboy-engine-primitives = { version = "0.1.0", path = "../homeboy-engine-primitives" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
